requirements.txt
appopener~=1.6
colorama~=0.4.6
feedparser~=6.0.11
googlesearch-python~=1.3.0
pillow~=11.1.0
pip-tools~=7.4.1
pipdeptree~=2.25.0
PyAudio~=0.2.14
PyAutoGUI~=0.9.54
pycaw~=20240210
python-dotenv~=1.0.1
pyttsx3~=2.98
regex~=2024.11.6
SpeechRecognition~=3.14.1
wikipedia~=1.4.0
WMI~=1.4.9
